fix: HA 2025.2.0 Compatability for new BT Devices
HA 2025.2.0 introduces new device entries in the core Bluetooth
integration for all HaScanners (like Shelly, esphome etc).
These changes account for that, and further allow the user to retain the
areas set on the esphome/shelly devices so as to not break existing
installs, or to assign areas using the BT device entries.

Signed-off-by: Ashley Gittins <ash@ajg.net.au>